✨ Connect Bottom Sheet
