--- MCDONALDS_iOS/MCDONALDS_iOS/OrderList/OrderListViewController.swift
@@ -1,56 +1,0 @@
-import UIKit
-import SnapKit
-import Then
-
-final class OrderListViewController: BaseViewController {
-    
-    private let rootView = OrderListView()
-    
-    override func loadView() {
-        view = rootView
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setRegister()
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-            super.viewWillAppear(animated)
-            setNavigationBar(type: .orderList)
-        }
-    
-    override func setAction() {
-        rootView.minusButton.addTarget(self, action: #selector(didTapMinus), for: .touchUpInside)
-        rootView.plusButton.addTarget(self, action: #selector(didTapPlus), for: .touchUpInside)
-        rootView.editButton.addTarget(self, action: #selector(didTapEdit), for: .touchUpInside)
-        rootView.addMenuButton.addTarget(self, action: #selector(didTapAddMenu), for: .touchUpInside)
-    }
-}
-
-// MARK: - Functions
-
-extension OrderListViewController {
-    private func setRegister() {
-        // Cell 등록 등
-    }
-    
-    @objc private func didTapMinus() {
-        // 수량 -1
-    }
-    
-    @objc private func didTapPlus() {
-        // 수량 +1
-    }
-    
-    @objc private func didTapEdit() {
-        // 메뉴 옵션 수정
-    }
-    
-    @objc private func didTapAddMenu() {
-        // + 메뉴 추가
-    }
-}
-
-
